Fix queue processing with an invalid email in it
This pull request fixes the following bug. If a customer inputs invalid characters in his emails address (for example, ī instead of i), his email address goes to the emails queue. When Magento cron starts to process this queue, an exception is being thrown, catched, and the whole process is stopped (there's a return statement in catch block). So no emails in queue, that are added after this invalid email address, won't be sent.
This pull request keeps the exception log, but marks the failed email as processed. So other customers will get their own emails.