Add idna-compatible mail-address encodings
With Zend_Mail and punyencoding there is the problem, that Zend_Mail is
not supporting the other. By encoding it in a IDN conform way it is
possible to use non-standard letters like german Umlauts, that are
allowed in domain-names but result in "5.1.3 Bad recipient address
syntax" if send to the server, even if they are UTF-8 encoded.